Enforced request content checking in back-end unit testing to ensure that required parameters are provided before accessing the database

+++ back-end/src/routes/deletePost.ts
@@ -7,14 +7,19 @@ const router = express.Router();
 // Delete API endpoint: deletes element from the database using provided ID
 router.delete('/:postId', authenticateToken, async (req, res) => {
     try {
-        const postId = req.params.postId;
-        const result = await pool.query('DELETE FROM blogs_table WHERE blog_id = $1', [postId]);
+        if (req.params.postId) {
+            const postId = req.params.postId;
+            const result = await pool.query('DELETE FROM blogs_table WHERE blog_id = $1', [postId]);
 
-        if (result.rowCount === 1) {
-            res.status(204).json(result.rows[0]);
-            console.log('Post deleted from the database:', postId);
-        } else {
-            res.status(404).json({ error: 'Post not found' });
+            if (result.rowCount === 1) {
+                res.status(204).json(result.rows[0]);
+                console.log('Post deleted from the database:', postId);
+            } else {
+                res.status(404).json({ error: 'Post not found' });
+            }
+        }
+        else {
+            res.status(404).json({ error: 'The request does not include a post ID' });
         }
     } catch (error) {
         console.error('Error deleting post from the database:', error);
